[Module] add dbo 3
[修改描述]
[问题描述]

+++ logic/yhc_file_storage.cpp
@@ -9,6 +9,7 @@
 
 
 FILE *FileStorage::m_pFileID[MAX_FILE_NUM];
+YHC_CHAR *FileStorage::m_pFilePath[MAX_FILE_NUM];
 
 
 FileStorage::FileStorage(char *pchrPath) : m_pchrPath(pchrPath)
@@ -57,28 +58,59 @@ YHC_S32 FileStorage::init(YHC_CHAR *pFileStr, YHC_U32 u32Id)
             return YHC_FAILURE;
         }
     }
+
     m_pFileID[u32Id] = pFile;
+    LOGW("m_pFileID[0]:%p", m_pFileID[0]);
 
     return YHC_SUCCESS;
 }
 
 YHC_S32 FileStorage::write(YHC_U32 u32pos, YHC_VOID *pData, YHC_U32 u32Length)
 {
-    YHC_S32 s32Ret = YHC_SUCCESS;
-    LOGI("aa");
+    if (YHC_NULL == m_pFileID[0])
+    {
+        LOGE("file not open");
+        return YHC_FAILURE;
+    }
 
-    fseek(m_pFileID[0], u32pos, SEEK_SET);
+    YHC_S32 s32Ret = YHC_SUCCESS;
+
+
+    LOGI("m_pFileID[0]:%p, u32pos:%d, ", m_pFileID[0], u32pos);
+
+    s32Ret = fseek(m_pFileID[0], u32pos, SEEK_SET);
+    if (s32Ret != YHC_SUCCESS)
+    {
+//        LOGE("fwrite ret:%s", perror("dddd"));
+        perror("dddd");
+    }
+
 
     LOGI("ab");
     s32Ret = fwrite(pData, 1, u32Length, m_pFileID[0]);
-    CHECK_RET_VALUE(s32Ret, "fwrite, u32Length:%d, ID:%d, u32pos:%d, ret:0x%x",
+    if (s32Ret != u32Length)
+    {
+    }
+    LOGE("fwrite, u32Length:%d, ID:%d, u32pos:%d, ret:0x%x",
         u32Length, 0, u32pos, s32Ret);
+
+    YHC_U32 DDDDD;
+    size(&DDDDD);
+    LOGF("size:%d", DDDDD);
+
     return YHC_SUCCESS;
 }
 
 
 YHC_S32 FileStorage::read(YHC_U32 u32pos, YHC_VOID *pData, YHC_U32 u32Length)
 {
+    if (YHC_NULL == m_pFileID[0])
+    {
+        LOGE("file not open");
+        return YHC_FAILURE;
+    }
+
+
     YHC_S32 s32ReadCnt = 0;
 
     fseek(m_pFileID[0], u32pos, SEEK_SET);
@@ -98,9 +130,15 @@ YHC_S32 FileStorage::read(YHC_U32 u32pos, YHC_VOID *pData, YHC_U32 u32Length)
 
 YHC_S32 FileStorage::size(YHC_U32 *pu32Size)
 {
+    if (YHC_NULL == m_pFileID[0])
+    {
+        LOGE("file not open");
+        return YHC_FAILURE;
+    }
 
     fseek(m_pFileID[0], 0, SEEK_END);
     *pu32Size = ftell(m_pFileID[0]);
+    LOGI("FileStorage size :%d", *pu32Size);
 
     return YHC_SUCCESS;
 
@@ -108,7 +146,20 @@ YHC_S32 FileStorage::size(YHC_U32 *pu32Size)
 
 YHC_S32 FileStorage::clear()
 {
+    if (YHC_NULL == m_pFileID[0])
+    {
+        LOGE("file not open");
+        return YHC_FAILURE;
+    }
     system("rm -rf /yhc/db/*");
+
+    m_pFileID[0] = fopen("/yhc/db/usr.db", "w+b");
+    if (YHC_NULL == m_pFileID[0])
+    {
+        LOGE("create file failure! ret:0x%x", errno);
+        return YHC_FAILURE;
+    }
+    LOGW("m_pFileID[0]:%p", m_pFileID[0]);
 
     return YHC_SUCCESS;
